+++ plugin/brdrq/brdrq_algorithm_autocorrectborders.py
@@ -31,15 +31,19 @@ import inspect
 import os
 import sys
 
+from qgis.core import QgsProcessingFeatureSourceDefinition
+
 from .brdrq_utils import (
     ENUM_REFERENCE_OPTIONS,
     ENUM_OD_STRATEGY_OPTIONS,
     ADPF_VERSIONS,
-    GRB_TYPES,
     geom_qgis_to_shapely,
     geojson_to_layer,
     get_workfolder,
     thematic_preparation,
+    get_symbol,
+    get_reference_params,
+    PREFIX_LOCAL_LAYER,
 )
 
 cmd_folder = os.path.split(inspect.getfile(inspect.currentframe()))[0]
@@ -130,9 +134,6 @@ class AutocorrectBordersProcessingAlgorithm(QgsProcessingAlgorithm):
 
     LAYER_RESULT_ACTUAL = "RESULT_ACTUAL"  # parameter that holds the TOC layername of the actualised result
     LAYER_RESULT_ACTUAL_DIFF = "RESULT_ACTUAL_DIFF"  # parameter that holds the TOC layername of the actualised resulting diff
-    PREFIX_LOCAL_LAYER = (
-        "LOCREF"  # prefix for the TOC layername, when a local layer is used
-    )
 
     # ALIGNER parameters
     CRS = "EPSG:31370"  # default CRS for the aligner,updated by CRS of thematic inputlayer
@@ -237,7 +238,7 @@ class AutocorrectBordersProcessingAlgorithm(QgsProcessingAlgorithm):
         parameter = QgsProcessingParameterFeatureSource(
             self.INPUT_THEMATIC,
             self.tr("THEMATIC LAYER"),
-            [QgsProcessing.TypeVectorPolygon],
+            [QgsProcessing.TypeVectorAnyGeometry],
             defaultValue="themelayer",
         )
         parameter.setFlags(parameter.flags())
@@ -263,8 +264,7 @@ class AutocorrectBordersProcessingAlgorithm(QgsProcessingAlgorithm):
         parameter = QgsProcessingParameterFeatureSource(
             self.INPUT_REFERENCE,
             self.tr("REFERENCE LAYER"),
-            [QgsProcessing.TypeVectorPolygon],
-            # defaultValue="referencelayer",
+            [QgsProcessing.TypeVectorAnyGeometry],
             optional=True,
         )
         parameter.setFlags(parameter.flags())
@@ -293,28 +293,28 @@ class AutocorrectBordersProcessingAlgorithm(QgsProcessingAlgorithm):
             QgsProcessingOutputVectorLayer(
                 "OUTPUT_RESULT",
                 self.LAYER_RESULT,
-                QgsProcessing.TypeVectorPolygon,
+                QgsProcessing.TypeVectorAnyGeometry,
             )
         )
         self.addOutput(
             QgsProcessingOutputVectorLayer(
                 "OUTPUT_RESULT_DIFF",
                 self.LAYER_RESULT_DIFF,
-                QgsProcessing.TypeVectorPolygon,
+                QgsProcessing.TypeVectorAnyGeometry,
             )
         )
         self.addOutput(
             QgsProcessingOutputVectorLayer(
                 "OUTPUT_RESULT_DIFF_PLUS",
                 self.LAYER_RESULT_DIFF_PLUS,
-                QgsProcessing.TypeVectorPolygon,
+                QgsProcessing.TypeVectorAnyGeometry,
             )
         )
         self.addOutput(
             QgsProcessingOutputVectorLayer(
                 "OUTPUT_RESULT_DIFF_MIN",
                 self.LAYER_RESULT_DIFF_MIN,
-                QgsProcessing.TypeVectorPolygon,
+                QgsProcessing.TypeVectorAnyGeometry,
             )
         )
         # advanced parameters
@@ -520,7 +520,7 @@ class AutocorrectBordersProcessingAlgorithm(QgsProcessingAlgorithm):
             aligner.load_reference_data(DictLoader(dict_reference))
             aligner.name_reference_id = self.ID_REFERENCE_FIELDNAME
             aligner.dict_reference_source["source"] = (
-                self.PREFIX_LOCAL_LAYER + "_" + self.LAYER_REFERENCE_NAME
+                PREFIX_LOCAL_LAYER + "_" + self.LAYER_REFERENCE_NAME
             )
             aligner.dict_reference_source["version_date"] = "unknown"
         elif self.SELECTED_REFERENCE in ADPF_VERSIONS:
@@ -624,10 +624,13 @@ class AutocorrectBordersProcessingAlgorithm(QgsProcessingAlgorithm):
 
         # MAKE TEMPORARY LAYERS
         if self.SELECTED_REFERENCE != 0:
+            reference_geojson = aligner.get_input_as_geojson(
+                inputtype=AlignerInputType.REFERENCE
+            )
             geojson_to_layer(
                 self.LAYER_REFERENCE_NAME,
-                aligner.get_input_as_geojson(inputtype=AlignerInputType.REFERENCE),
-                QgsStyle.defaultStyle().symbol("outline black"),
+                reference_geojson,
+                get_symbol(reference_geojson, "reference"),
                 True,
                 self.GROUP_LAYER,
                 self.WORKFOLDER,
@@ -652,35 +655,42 @@ class AutocorrectBordersProcessingAlgorithm(QgsProcessingAlgorithm):
                     self.GROUP_LAYER,
                     self.WORKFOLDER,
                 )
-
+        result_diff = "result_diff"
+        geojson_result_diff = fcs[result_diff]
         geojson_to_layer(
             self.LAYER_RESULT_DIFF,
-            fcs["result_diff"],
-            QgsStyle.defaultStyle().symbol("hashed black X"),
+            geojson_result_diff,
+            get_symbol(geojson_result_diff, result_diff),
             False,
             self.GROUP_LAYER,
             self.WORKFOLDER,
         )
+        result_diff_plus = "result_diff_plus"
+        geojson_result_diff_plus = fcs[result_diff_plus]
         geojson_to_layer(
             self.LAYER_RESULT_DIFF_PLUS,
-            fcs["result_diff_plus"],
-            QgsStyle.defaultStyle().symbol("hashed cgreen /"),
+            geojson_result_diff_plus,
+            get_symbol(geojson_result_diff_plus, result_diff_plus),
             False,
             self.GROUP_LAYER,
             self.WORKFOLDER,
         )
+        result_diff_min = "result_diff_min"
+        geojson_result_diff_min = fcs[result_diff_min]
         geojson_to_layer(
             self.LAYER_RESULT_DIFF_MIN,
-            fcs["result_diff_min"],
-            QgsStyle.defaultStyle().symbol("hashed cred /"),
+            geojson_result_diff_min,
+            get_symbol(geojson_result_diff_min, result_diff_min),
             False,
             self.GROUP_LAYER,
             self.WORKFOLDER,
         )
+        result = "result"
+        geojson_result = fcs[result]
         geojson_to_layer(
             self.LAYER_RESULT,
-            fcs["result"],
-            QgsStyle.defaultStyle().symbol("outline green"),
+            geojson_result,
+            get_symbol(geojson_result, result),
             True,
             self.GROUP_LAYER,
             self.WORKFOLDER,
@@ -772,9 +782,17 @@ class AutocorrectBordersProcessingAlgorithm(QgsProcessingAlgorithm):
             wrkfldr, name="autocorrectborders", temporary=False
         )
         self.RELEVANT_DISTANCE = parameters["RELEVANT_DISTANCE"]
-        self.LAYER_THEMATIC = self.parameterAsVectorLayer(
-            parameters, self.INPUT_THEMATIC, context
-        )
+        param_input_thematic = parameters[self.INPUT_THEMATIC]
+        if isinstance(
+            parameters[self.INPUT_THEMATIC], QgsProcessingFeatureSourceDefinition
+        ):
+            self.LAYER_THEMATIC = QgsProject.instance().mapLayer(
+                param_input_thematic.toVariant()["source"]["val"]
+            )
+        else:
+            self.LAYER_THEMATIC = self.parameterAsVectorLayer(
+                parameters, self.INPUT_THEMATIC, context
+            )
         self.CRS = (
             self.LAYER_THEMATIC.sourceCrs().authid()
         )  # set CRS for the calculations, based on the THEMATIC input layer
@@ -800,34 +818,15 @@ class AutocorrectBordersProcessingAlgorithm(QgsProcessingAlgorithm):
         self.SHOW_LOG_INFO = parameters["SHOW_LOG_INFO"]
 
         ref = ENUM_REFERENCE_OPTIONS[parameters["ENUM_REFERENCE"]]
-
-        if ref in GRB_TYPES:
-            self.SELECTED_REFERENCE = GRBType[ref]
-            self.LAYER_REFERENCE_NAME = GRBType[ref]
-            ref_suffix = str(ref)
-        elif ref in ADPF_VERSIONS:
-            self.SELECTED_REFERENCE = ref
-            self.LAYER_REFERENCE_NAME = ref
-            ref_suffix = str(ref)
-        else:
-            self.SELECTED_REFERENCE = 0
-            self.LAYER_REFERENCE = self.parameterAsVectorLayer(
-                parameters, self.INPUT_REFERENCE, context
+        self.LAYER_REFERENCE = self.parameterAsVectorLayer(
+            parameters, self.INPUT_REFERENCE, context
+        )
+        self.SELECTED_REFERENCE, self.LAYER_REFERENCE_NAME, ref_suffix = (
+            get_reference_params(
+                ref, self.LAYER_REFERENCE, self.ID_REFERENCE_FIELDNAME, self.CRS
             )
+        )
 
-            if self.LAYER_REFERENCE is None or self.ID_REFERENCE_FIELDNAME == "NULL":
-                raise QgsProcessingException(
-                    "Please choose a REFERENCELAYER from the table of contents, and the associated unique REFERENCE ID"
-                )
-
-            self.LAYER_REFERENCE_NAME = self.LAYER_REFERENCE.name()
-            ref_suffix = self.PREFIX_LOCAL_LAYER + "_" + self.LAYER_REFERENCE_NAME
-
-            if self.LAYER_REFERENCE.sourceCrs().authid() != self.CRS:
-                raise QgsProcessingException(
-                    "Thematic layer and ReferenceLayer are in a different CRS. "
-                    "Please provide them in the same CRS, with units in meter (f.e. For Belgium in EPSG:31370 or EPSG:3812)"
-                )
         self.SUFFIX = "_DIST_" + str(self.RELEVANT_DISTANCE) + "_" + ref_suffix
         if self.PREDICTIONS:
             self.SUFFIX = self.SUFFIX + "_PREDICTIONS"