+++ plugin/brdrq/brdrq_algorithm_autoupdateborders.py
@@ -31,7 +31,7 @@ THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
 from brdr.aligner import Aligner
 from brdr.constants import BASE_FORMULA_FIELD_NAME
-from brdr.enums import AlignerInputType, GRBType, FullStrategy
+from brdr.enums import AlignerInputType, GRBType, FullStrategy, OpenDomainStrategy
 from brdr.grb import update_to_actual_grb
 from brdr.loader import DictLoader
 from qgis.PyQt.QtCore import QCoreApplication
@@ -39,6 +39,7 @@ from qgis.PyQt.QtCore import QDate, QDateTime
 from qgis.core import QgsProcessing
 from qgis.core import QgsProcessingAlgorithm
 from qgis.core import QgsProcessingException
+from qgis.core import QgsProcessingFeatureSourceDefinition
 from qgis.core import QgsProcessingMultiStepFeedback
 from qgis.core import QgsProcessingOutputVectorLayer
 from qgis.core import QgsProcessingParameterBoolean
@@ -50,7 +51,6 @@ from qgis.core import (
 )
 from qgis.core import QgsProcessingParameterFile
 from qgis.core import QgsProject
-from qgis.core import QgsStyle
 
 from .brdrq_utils import (
     geom_qgis_to_shapely,
@@ -61,6 +61,9 @@ from .brdrq_utils import (
     ENUM_PREDICTION_STRATEGY_OPTIONS,
     PredictionStrategy,
     ENUM_FULL_STRATEGY_OPTIONS,
+    ENUM_OD_STRATEGY_OPTIONS,
+    get_symbol,
+    get_reference_params,
 )
 
 
@@ -203,7 +206,7 @@ class AutoUpdateBordersProcessingAlgorithm(QgsProcessingAlgorithm):
         parameter = QgsProcessingParameterFeatureSource(
             self.INPUT_THEMATIC,
             "THEMATIC LAYER, with the features to align",
-            [QgsProcessing.TypeVectorPolygon],
+            [QgsProcessing.TypeVectorAnyGeometry],
             defaultValue="themelayer",
         )
         parameter.setFlags(parameter.flags())
@@ -252,6 +255,28 @@ class AutoUpdateBordersProcessingAlgorithm(QgsProcessingAlgorithm):
 
         # ADVANCED INPUT
         parameter = QgsProcessingParameterEnum(
+            "ENUM_OD_STRATEGY",
+            "Select OD-STRATEGY:",
+            options=ENUM_OD_STRATEGY_OPTIONS,
+            defaultValue=3,  # Index of the default option (e.g., 'SNAP_ALL_SIDE')
+        )
+        parameter.setFlags(
+            parameter.flags() | QgsProcessingParameterDefinition.FlagAdvanced
+        )
+        self.addParameter(parameter)
+
+        parameter = QgsProcessingParameterNumber(
+            "THRESHOLD_OVERLAP_PERCENTAGE",
+            "THRESHOLD_OVERLAP_PERCENTAGE (%)",
+            type=QgsProcessingParameterNumber.Double,
+            defaultValue=50,
+        )
+        parameter.setFlags(
+            parameter.flags() | QgsProcessingParameterDefinition.FlagAdvanced
+        )
+        self.addParameter(parameter)
+
+        parameter = QgsProcessingParameterEnum(
             "FULL_STRATEGY",
             "Select FULL_STRATEGY:",
             options=ENUM_FULL_STRATEGY_OPTIONS,
@@ -299,28 +324,28 @@ class AutoUpdateBordersProcessingAlgorithm(QgsProcessingAlgorithm):
             QgsProcessingOutputVectorLayer(
                 "OUTPUT_RESULT",
                 self.LAYER_RESULT,
-                QgsProcessing.TypeVectorPolygon,
+                QgsProcessing.TypeVectorAnyGeometry,
             )
         )
         self.addOutput(
             QgsProcessingOutputVectorLayer(
                 "OUTPUT_RESULT_DIFF",
                 self.LAYER_RESULT_DIFF,
-                QgsProcessing.TypeVectorPolygon,
+                QgsProcessing.TypeVectorAnyGeometry,
             )
         )
         self.addOutput(
             QgsProcessingOutputVectorLayer(
                 "OUTPUT_RESULT_DIFF_PLUS",
                 self.LAYER_RESULT_DIFF_PLUS,
-                QgsProcessing.TypeVectorPolygon,
+                QgsProcessing.TypeVectorAnyGeometry,
             )
         )
         self.addOutput(
             QgsProcessingOutputVectorLayer(
                 "OUTPUT_RESULT_DIFF_MIN",
                 self.LAYER_RESULT_DIFF_MIN,
-                QgsProcessing.TypeVectorPolygon,
+                QgsProcessing.TypeVectorAnyGeometry,
             )
         )
 
@@ -369,7 +394,10 @@ class AutoUpdateBordersProcessingAlgorithm(QgsProcessingAlgorithm):
                     attributes_dict[key] = value
             dict_thematic_properties[id_theme] = attributes_dict
 
-        aligner = Aligner(od_strategy=self.OD_STRATEGY)
+        aligner = Aligner(
+            od_strategy=self.OD_STRATEGY,
+            threshold_overlap_percentage=self.THRESHOLD_OVERLAP_PERCENTAGE,
+        )
         aligner.load_thematic_data(
             DictLoader(
                 data_dict=dict_thematic, data_dict_properties=dict_thematic_properties
@@ -415,40 +443,50 @@ class AutoUpdateBordersProcessingAlgorithm(QgsProcessingAlgorithm):
 
         # Add RESULT TO TOC
         if "result_diff_min" in fcs_actualisation:
+            result_diff_min = "result_diff_min"
+            geojson_result_diff_min = fcs_actualisation[result_diff_min]
             geojson_to_layer(
                 self.LAYER_RESULT_DIFF_MIN,
-                fcs_actualisation["result_diff_min"],
-                QgsStyle.defaultStyle().symbol("hashed cred /"),
+                geojson_result_diff_min,
+                get_symbol(geojson_result_diff_min, result_diff_min),
                 True,
                 self.GROUP_LAYER,
                 self.WORKFOLDER,
             )
         if "result_diff_plus" in fcs_actualisation:
+            result_diff_plus = "result_diff_plus"
+            geojson_result_diff_plus = fcs_actualisation[result_diff_plus]
             geojson_to_layer(
                 self.LAYER_RESULT_DIFF_PLUS,
-                fcs_actualisation["result_diff_plus"],
-                QgsStyle.defaultStyle().symbol("gradient green fill"),
+                geojson_result_diff_plus,
+                get_symbol(geojson_result_diff_plus, result_diff_plus),
                 True,
                 self.GROUP_LAYER,
                 self.WORKFOLDER,
             )
         if "result_diff" in fcs_actualisation:
+            result_diff = "result_diff"
+            geojson_result_diff = fcs_actualisation[result_diff]
             geojson_to_layer(
                 self.LAYER_RESULT_DIFF,
-                fcs_actualisation["result_diff"],
-                QgsStyle.defaultStyle().symbol("hashed black X"),
+                geojson_result_diff,
+                get_symbol(geojson_result_diff, result_diff),
                 False,
                 self.GROUP_LAYER,
                 self.WORKFOLDER,
             )
+
+        result = "result"
+        geojson_result = fcs_actualisation[result]
         geojson_to_layer(
             self.LAYER_RESULT,
-            fcs_actualisation["result"],
-            QgsStyle.defaultStyle().symbol("outline green"),
+            geojson_result,
+            get_symbol(geojson_result, result),
             True,
             self.GROUP_LAYER,
             self.WORKFOLDER,
         )
+
         feedback.pushInfo("Resulterende geometrie berekend")
         feedback.pushInfo("END ACTUALISATION")
         result = QgsProject.instance().mapLayersByName(self.LAYER_RESULT)[0]
@@ -481,14 +519,29 @@ class AutoUpdateBordersProcessingAlgorithm(QgsProcessingAlgorithm):
         self.WORKFOLDER = get_workfolder(
             wrkfldr, name="autoupdateborders", temporary=False
         )
-        self.THEMATIC_LAYER = self.parameterAsVectorLayer(
-            parameters, self.INPUT_THEMATIC, context
-        )
+        param_input_thematic = parameters[self.INPUT_THEMATIC]
+        if isinstance(
+            parameters[self.INPUT_THEMATIC], QgsProcessingFeatureSourceDefinition
+        ):
+            self.THEMATIC_LAYER = QgsProject.instance().mapLayer(
+                param_input_thematic.toVariant()["source"]["val"]
+            )
+        else:
+            self.THEMATIC_LAYER = self.parameterAsVectorLayer(
+                parameters, self.INPUT_THEMATIC, context
+            )
         self.CRS = (
             self.THEMATIC_LAYER.sourceCrs().authid()
         )  # set CRS for the calculations, based on the THEMATIC input layer
         self.MAX_DISTANCE_FOR_ACTUALISATION = parameters["MAX_RELEVANT_DISTANCE"]
-        self.GRB_TYPE = GRBType[GRB_TYPES[parameters["ENUM_REFERENCE"]]]
+        self.THRESHOLD_OVERLAP_PERCENTAGE = parameters["THRESHOLD_OVERLAP_PERCENTAGE"]
+        self.OD_STRATEGY = OpenDomainStrategy[
+            ENUM_OD_STRATEGY_OPTIONS[parameters["ENUM_OD_STRATEGY"]]
+        ]
+        ref = GRB_TYPES[parameters["ENUM_REFERENCE"]]
+        self.GRB_TYPE, layer_reference_name, ref_suffix = get_reference_params(
+            ref, None, None, self.CRS
+        )
         self.SHOW_LOG_INFO = parameters["SHOW_LOG_INFO"]
         self.PREDICTION_STRATEGY = PredictionStrategy[
             ENUM_PREDICTION_STRATEGY_OPTIONS[parameters["PREDICTION_STRATEGY"]]