+++ plugin/brdrq/brdrq_utils.py
@@ -1,6 +1,8 @@
 import os
 from enum import Enum
 
+from qgis.core import QgsProcessingException
+
 try:
     import brdr
 except:
@@ -41,6 +43,7 @@ from qgis.utils import iface
 from shapely import to_wkt, from_wkt, make_valid
 
 
+# TODO remove enum class when available in brdr (v0.12.0)
 class PredictionStrategy(str, Enum):
     """
     Enum for prediction strategy when using GRB updater
@@ -55,22 +58,34 @@ class PredictionStrategy(str, Enum):
     ORIGINAL = "original"
 
 
-LOCAL_REFERENCE_LAYER = "LOCAL REFERENCE LAYER (choose LAYER and ID below)"
+SPLITTER = ":"
+PREFIX_LOCAL_LAYER = (
+    "LOCREF"  # prefix for the TOC layername, when a local layer is used
+)
+LOCAL_REFERENCE_LAYER = (
+    PREFIX_LOCAL_LAYER + SPLITTER + " choose LOCAL LAYER and UNIQUE ID below"
+)
 
-GRB_TYPES = [
-    e.name for e in GRBType
-]  # types of actual GRB: parcels, buildings, artwork
-ADPF_VERSIONS = [
-    "Adpf" + str(x) for x in [datetime.datetime.today().year - i for i in range(6)]
-]  # Fiscal parcels of past 5 years
+DICT_REFERENCE_OPTIONS = dict()
+DICT_REFERENCE_OPTIONS[LOCAL_REFERENCE_LAYER] = PREFIX_LOCAL_LAYER
 
-ENUM_REFERENCE_OPTIONS = (
-    [LOCAL_REFERENCE_LAYER] + GRB_TYPES + ADPF_VERSIONS
-)  # Options for downloadable reference layers
+DICT_GRB_TYPES = dict()
+for e in GRBType:
+    DICT_GRB_TYPES[e.name + SPLITTER + " GRB " + e.value.split(" - ")[2]] = e.name
+DICT_ADPF_VERSIONS = dict()
+for x in [datetime.datetime.today().year - i for i in range(6)]:
+    DICT_ADPF_VERSIONS["Administratieve fiscale percelen" + SPLITTER + " " + str(x)] = x
+
+DICT_REFERENCE_OPTIONS.update(DICT_GRB_TYPES)
+DICT_REFERENCE_OPTIONS.update(DICT_ADPF_VERSIONS)
+
+GRB_TYPES = list(DICT_GRB_TYPES.keys())
+ADPF_VERSIONS = list(DICT_ADPF_VERSIONS.keys())
+ENUM_REFERENCE_OPTIONS = list(DICT_REFERENCE_OPTIONS.keys())
 
 # ENUM for choosing the OD-strategy
-ENUM_OD_STRATEGY_OPTIONS = [
-    e.name for e in OpenDomainStrategy if e.value <= 2
+ENUM_OD_STRATEGY_OPTIONS = [e.name for e in OpenDomainStrategy][
+    :4
 ]  # list with od-strategy-options #if e.value<=2
 
 # ENUM for choosing the snap-strategy
@@ -251,6 +266,8 @@ def get_symbol(geojson, resulttype):
             return QgsStyle.defaultStyle().symbol("hashed cred /")
         elif resulttype == "result":
             return QgsStyle.defaultStyle().symbol("outline green")
+        elif resulttype == "reference":
+            return QgsStyle.defaultStyle().symbol("outline black")
         else:
             return QgsStyle.defaultStyle().symbol("outline blue")
     elif geometrytype == "MultiLineString":
@@ -262,6 +279,8 @@ def get_symbol(geojson, resulttype):
             return QgsStyle.defaultStyle().symbol("dash red")
         elif resulttype == "result":
             return QgsStyle.defaultStyle().symbol("simple green line")
+        elif resulttype == "reference":
+            return QgsStyle.defaultStyle().symbol("simple black line")
         else:
             return QgsStyle.defaultStyle().symbol("simple blue line")
     elif geometrytype == "MultiPoint":
@@ -273,6 +292,8 @@ def get_symbol(geojson, resulttype):
             return QgsStyle.defaultStyle().symbol("dot white")
         elif resulttype == "result":
             return QgsStyle.defaultStyle().symbol("dot green")
+        elif resulttype == "reference":
+            return QgsStyle.defaultStyle().symbol("dot black")
         else:
             return QgsStyle.defaultStyle().symbol("dot blue")
     else:
@@ -575,6 +596,34 @@ def _processresult_to_dicts(processresult):
         results_relevant_intersection,
         results_relevant_diff,
     )
+
+
+def get_reference_params(ref, layer_reference, id_reference_fieldname, thematic_crs):
+
+    ref_id = DICT_REFERENCE_OPTIONS[ref]
+    if ref in GRB_TYPES:
+        selected_reference = GRBType[ref_id]
+        layer_reference_name = GRBType[ref_id].name
+        ref_suffix = str(ref_id)
+        print(selected_reference)
+    elif ref in ADPF_VERSIONS:
+        selected_reference = ref_id
+        layer_reference_name = ref_id
+        ref_suffix = str(ref_id)
+    else:
+        selected_reference = 0
+        if layer_reference is None or id_reference_fieldname == "NULL":
+            raise QgsProcessingException(
+                "Please choose a REFERENCELAYER from the table of contents, and the associated unique REFERENCE ID"
+            )
+        layer_reference_name = layer_reference.name()
+        ref_suffix = PREFIX_LOCAL_LAYER + "_" + layer_reference_name
+        if layer_reference.sourceCrs().authid() != thematic_crs:
+            raise QgsProcessingException(
+                "Thematic layer and ReferenceLayer are in a different CRS. "
+                "Please provide them in the same CRS, with units in meter (f.e. For Belgium in EPSG:31370 or EPSG:3812)"
+            )
+    return selected_reference, layer_reference_name, ref_suffix
 
 
 def thematic_preparation(input_thematic_layer, relevant_distance, context, feedback):